Reset button has been added
The reset button clears the canvas so that the user can start drawing
again.

It works by deleting all the child elements and adding them all back in
as shown by the resetCanvas() button.

The faster and more efficient way of doing it would be to change all the
dots on the canvas to a white background again but that isn't working in
the moment.