Color picker has been added
Color picker has been added, by default the mouse pointer will turn the
dots of the canvas to a random color; but when selecting a color form
the color picker, the color of each dot changes to what's within the
color picker.

Next step is to add a button that changes back to random colors and
oraganize the code.