New slider and some comments deleted
On top of the button div, slider has been added along with a label that
tells of it's value.

Currently, there's no functionality to it as I need to make adjustments
to how the canvas behaves first.

Some comments were deleted too since they didn't provide information
that wasn't shown within the code.