Preparing for minecraft theme
Bruh, I just forgot to create a new branch before making these changes
so now i gotta commit to main before I start testing.

Anyways, the UI is more organized to better suit the look I'm going for.
To elaborate, the size change slider along with brush and background
color picker will be set to left side while the canvas and buttons are
planned to go to the right. Header will by on top and centered.