Images and prototype of layout created
Images for the brush and background tool have been created. Later on in
the project, the icons will replace the text of the color pickers so
users get a visual look on what each thing does.

The layout of the website now has canvas as the centerpeice with
controls on the left and right of it. It doesn't wrap around as well as
I would like but that can be fixed.