Touch and drag works similar to Mouse
Now there's an actionlistener on the canvasDiv that allows for painting
the canvas just like using a mouse... EXCEPT if you hover your finger
over the same point, the color continues to change.

I've tried mutliple solutions (giving each dot a boolean attribute that
stated if the dot has already been colored, trying to make the
colorManager line activate only once, and trying to make the actionlistener
understand if it's on the same dot or not). I'm not sure how to fix it
right now but I think for the time being it's not going to be a
priority.