Fixed Bugs with rainbow mode on and 1x1 canvas
Originally, when rainbow mode was on the color of the selected dot would
change rapidly if the finger were to hover over it. Now it's fixed
useing some extra functions to go along with touchmove.

Also, the 1x1 canvas would never change colors on touch and desktop. Now
that's fixed after allowing each dot to draw as soon as the mouse/touch
is starts. Which also makes the canvas feel more responsive.

Code has also been refractored to be more constistent and consice. What
needs to be done next is some explanations over what some of the more
complicated functions do.