Added mc-background, buttons, text, & slider
The background and music for the website has been set to a loop just like the
minecraft java main menu. Text has also been changed to white with a
pixelated shadow and in the minecraft theme.

Buttons and slider similar to the game have also been
added but they don't look as accurate to the game as of 2023.

Next theme modification: Frame, fix slider, and color pickers.